backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
autogen==0.3.1
bcrypt==4.2.0
beanie==1.27.0
biopython==1.84
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==43.0.3
debugpy==1.8.6
decorator==5.1.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
ecdsa==0.19.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.4
fastapi-cli==0.0.5
FLAML==2.3.2
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.7
langchain-core==0.3.15
langchain-google-genai==2.0.4
langchain-text-splitters==0.3.2
langsmith==0.1.140
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
motor==3.6.0
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.54.3
orjson==3.10.11
packaging==24.1
parso==0.8.4
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.9.2
pyotp==2.9.0
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
pywin32==307
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.2
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
wheel==0.44.0
yarl==1.17.1
